made and implemented (when loading main game) animated load screen.
made all floor mats emissive. (there was an issue with lighting)

created a basic reset which is used when you crash on the corner, resetting to when you enter the corner instead of after you've cornered.

noticed i kept hitting the hay cart so have made it very slightly thinner on one of the axis. (X i think)